logistics-assistant-backend/requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-community
sseclient-py
python-dotenv
